mcpserver/requirements.txt
mcp>=1.0.0
pydantic>=2.0.0
google-cloud-firestore[async]>=2.16.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
